restructured the code and removed code_helpers.py
